requirements.txt
# WattsTap Referral Service - Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Authentication
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

# HTTP Client (for Telegram API)
httpx==0.26.0

# Environment
python-dotenv==1.0.1

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0



